+++ packages/davecode-types/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @davecode/types
 
+## 1.0.1
+
+### Patch Changes
+
+- Immutable, Mutable, and PreventAutoSimplify types
+- 41c1e45: add Immutable<T>
+
 ## 1.0.0
 
 ### Major Changes