+++ packages/eslint-config-dave/CHANGELOG.md
@@ -1,5 +1,11 @@
 # eslint-config-dave
 
+## 3.0.1
+
+### Patch Changes
+
+- 89a90cb: disable 'curly' rule
+
 ## 3.0.0
 
 ### Major Changes